+++ packages/form/src/layouts/QueryFilter/style.ts
@@ -13,8 +13,21 @@ const genProStyle: GenerateStyle<ProToken> = (token) => {
         [`${token.antCls}-form-item`]: {
           marginBlock: 0,
         },
+        [`${token.proComponentsCls}-form-group-title`]: {
+          marginBlock: 0,
+        },
         '&-row': {
           rowGap: 24,
+          '&-split-line': {
+            '&:after': {
+              position: 'absolute',
+              width: '100%',
+              content: '""',
+              height: 1,
+              insetBlockEnd: -12,
+              borderBlockEnd: '1px dashed ' + token.colorSplit,
+            },
+          },
         },
         '&-collapse-button': {
           display: 'flex',