+++ packages/layout/src/components/AppsLogoComponents/index.tsx
@@ -39,7 +39,7 @@ export const AppsLogoComponents: React.FC<{
   const { appList, prefixCls = 'ant-pro' } = props;
   const ref = React.useRef<HTMLDivElement>(null);
   const baseClassName = `${prefixCls}-layout-apps`;
-  const { wrapSSR } = useStyle(baseClassName);
+  const { wrapSSR, hashId } = useStyle(baseClassName);
 
   const [visible, setVisible] = useState(false);
 
@@ -48,10 +48,22 @@ export const AppsLogoComponents: React.FC<{
       return !app?.desc;
     });
     if (isSimple) {
-      return <SimpleContent appList={appList} baseClassName={`${baseClassName}-simple`} />;
+      return (
+        <SimpleContent
+          hashId={hashId}
+          appList={appList}
+          baseClassName={`${baseClassName}-simple`}
+        />
+      );
     }
-    return <DefaultContent appList={appList} baseClassName={`${baseClassName}-default`} />;
-  }, [appList, baseClassName]);
+    return (
+      <DefaultContent
+        hashId={hashId}
+        appList={appList}
+        baseClassName={`${baseClassName}-default`}
+      />
+    );
+  }, [appList, baseClassName, hashId]);
 
   if (!props?.appList?.length) return null;
 
@@ -64,7 +76,7 @@ export const AppsLogoComponents: React.FC<{
         zIndex={9999}
         arrowPointAtCenter
         onVisibleChange={setVisible}
-        overlayClassName={`${baseClassName}-popover`}
+        overlayClassName={`${baseClassName}-popover ${hashId}`}
         content={popoverContent}
         getPopupContainer={() => ref.current || document.body}
       >
@@ -72,7 +84,7 @@ export const AppsLogoComponents: React.FC<{
           onClick={(e) => {
             e.stopPropagation();
           }}
-          className={classNames(`${baseClassName}-icon`, {
+          className={classNames(`${baseClassName}-icon`, hashId, {
             [`${baseClassName}-icon-active`]: visible,
           })}
         >