+++ packages/layout/src/components/Header.tsx
@@ -89,7 +89,9 @@ const DefaultHeader: React.FC<HeaderViewProps & PrivateSiderMenuProps> = (props)
   return wrapSSR(
     <>
       <ConfigProvider
+        // @ts-ignore
         theme={{
+          hashed: process.env.NODE_ENV?.toLowerCase() !== 'test',
           override: {
             Layout: {
               colorBgHeader: 'transparent',