+++ packages/layout/src/components/SiderMenu/BaseMenu.tsx
@@ -93,14 +93,13 @@ let IconFont = createFromIconfontCN({
 //   icon: '/favicon.png',
 //   icon: <Icon type="setting" />,
 const getIcon = (
-  icon?: string | React.ReactNode,
+  icon: string | React.ReactNode,
   iconPrefixes: string = 'icon-',
+  className: string,
 ): React.ReactNode => {
   if (typeof icon === 'string' && icon !== '') {
     if (isUrl(icon) || isImg(icon)) {
-      return (
-        <Icon component={() => <img src={icon} alt="icon" className="ant-pro-sider-menu-icon" />} />
-      );
+      return <Icon component={() => <img src={icon} alt="icon" className={className} />} />;
     }
     if (icon.startsWith(iconPrefixes)) {
       return <IconFont type={icon} />;
@@ -122,6 +121,7 @@ class MenuUtil {
       token?: LayoutDesignToken;
       menuRenderType?: 'header' | 'sider';
       baseClassName: string;
+      hashId: string;
     },
   ) {
     this.props = props;
@@ -131,6 +131,7 @@ class MenuUtil {
     token?: LayoutDesignToken;
     menuRenderType?: 'header' | 'sider';
     baseClassName: string;
+    hashId: string;
   };
 
   getNavMenuItems = (menusData: MenuDataItem[] = [], level: number): ItemType[] =>
@@ -155,7 +156,11 @@ class MenuUtil {
       const shouldHasIcon = level === 0 || (isGroup && level === 1);
 
       //  get defaultTitle by menuItemRender
-      const iconDom = getIcon(item.icon, iconPrefixes);
+      const iconDom = getIcon(
+        item.icon,
+        iconPrefixes,
+        `${baseClassName}-icon ${this.props?.hashId}`,
+      );
       /**
        * 如果没有icon在收起的时候用首字母代替
        */
@@ -164,19 +169,21 @@ class MenuUtil {
       const defaultTitle = (
         <div
           title={name}
-          className={classNames(`${baseClassName}-item-title`, {
+          className={classNames(`${baseClassName}-item-title`, this.props?.hashId, {
             [`${baseClassName}-item-title-collapsed`]: collapsed,
             [`${baseClassName}-item-collapsed-show-title`]: menu?.collapsedShowTitle && collapsed,
           })}
         >
           {/* 收起的时候group模式就不要展示icon了，放不下 */}
           {menuType === 'group' && collapsed ? null : shouldHasIcon && iconDom ? (
-            <span className={`anticon ${baseClassName}-item-icon`}>{iconDom}</span>
+            <span className={`anticon ${baseClassName}-item-icon ${this.props?.hashId}`}>
+              {iconDom}
+            </span>
           ) : (
             defaultIcon
           )}
           <span
-            className={classNames(`${baseClassName}-item-text`, {
+            className={classNames(`${baseClassName}-item-text`, this.props?.hashId, {
               [`${baseClassName}-item-text-has-icon`]:
                 menuType !== 'group' && shouldHasIcon && (iconDom || defaultIcon),
             })}
@@ -213,7 +220,7 @@ class MenuUtil {
               key: (item.key! || item.path!) + '-group-divider',
               style: {
                 padding: 0,
-                borderBottom: 0,
+                borderBlockEnd: 0,
                 margin: this.props.collapsed ? '4px' : '6px 16px',
                 marginBlockStart: this.props.collapsed ? 4 : 8,
                 borderColor: designToken?.sider?.colorMenuItemDivider,
@@ -266,18 +273,24 @@ class MenuUtil {
     const isGroup = menu?.type === 'group';
     /** Menu 第一级可以有icon，或者 isGroup 时第二级别也要有 */
     const hasIcon = level === 0 || (isGroup && level === 1);
-    const icon = !hasIcon ? null : getIcon(item.icon, iconPrefixes);
+    const icon = !hasIcon
+      ? null
+      : getIcon(item.icon, iconPrefixes, `${baseClassName}-icon ${this.props?.hashId}`);
     const defaultIcon = collapsed && hasIcon ? getMenuTitleSymbol(name) : null;
     let defaultItem = (
       <div
-        className={classNames(`${baseClassName}-item-title`, {
+        className={classNames(`${baseClassName}-item-title`, this.props?.hashId, {
           [`${baseClassName}-item-title-collapsed`]: collapsed,
           [`${baseClassName}-item-collapsed-show-title`]: menu?.collapsedShowTitle && collapsed,
         })}
       >
-        {icon ? <span className={`anticon ${baseClassName}-item-icon`}>{icon}</span> : defaultIcon}
+        {icon ? (
+          <span className={`anticon ${baseClassName}-item-icon ${this.props?.hashId}`}>{icon}</span>
+        ) : (
+          defaultIcon
+        )}
         <span
-          className={classNames(`${baseClassName}-item-text`, {
+          className={classNames(`${baseClassName}-item-text`, this.props?.hashId, {
             [`${baseClassName}-item-text-has-icon`]: hasIcon && (icon || defaultIcon),
           })}
         >
@@ -295,19 +308,21 @@ class MenuUtil {
           onClick={() => {
             window?.open?.(itemPath, '_blank');
           }}
-          className={classNames(`${baseClassName}-item-title`, {
+          className={classNames(`${baseClassName}-item-title`, this.props?.hashId, {
             [`${baseClassName}-item-title-collapsed`]: collapsed,
             [`${baseClassName}-item-link`]: true,
             [`${baseClassName}-item-collapsed-show-title`]: menu?.collapsedShowTitle && collapsed,
           })}
         >
           {icon ? (
-            <span className={`anticon ${baseClassName}-item-icon`}>{icon}</span>
+            <span className={`anticon ${baseClassName}-item-icon ${this.props?.hashId}`}>
+              {icon}
+            </span>
           ) : (
             defaultIcon
           )}
           <span
-            className={classNames(`${baseClassName}-item-text`, {
+            className={classNames(`${baseClassName}-item-text`, this.props?.hashId, {
               [`${baseClassName}-item-text-has-icon`]: hasIcon && (icon || defaultIcon),
             })}
           >
@@ -462,6 +477,7 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [openKeys && openKeys.join(','), props.layout, props.collapsed],
   );
+  const { wrapSSR, hashId } = useStyle(baseClassName);
 
   const menuUtils = useMemo(() => {
     return new MenuUtil({
@@ -469,10 +485,9 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
       token: designToken,
       menuRenderType,
       baseClassName,
+      hashId,
     });
-  }, [designToken, baseClassName, menuRenderType, props]);
-
-  const { wrapSSR, hashId } = useStyle(baseClassName);
+  }, [props, designToken, menuRenderType, baseClassName, hashId]);
 
   if (menu?.loading) {
     return (