+++ packages/layout/src/demos/siteMenu.tsx
@@ -199,7 +199,7 @@ export default () => {
                       style={{
                         padding: '32px 40px',
                         backgroundColor: '#fff',
-                        width: '100vw',
+                        width: 'calc(100vw - 4px)',
                         height: '307px',
                         boxShadow:
                           '0 8px 16px 0 rgba(0,0,0,0.03), 0 4px 8px 0 rgba(25,15,15,0.07), 0 2px 4px 0 rgba(0,0,0,0.08)',
@@ -220,7 +220,7 @@ export default () => {
                         <div
                           style={{
                             width: '308px',
-                            borderLeft: '1px solid rgba(0,0,0,0.06)',
+                            borderInlineStart: '1px solid rgba(0,0,0,0.06)',
                             paddingInlineStart: 16,
                           }}
                         >