+++ packages/utils/src/useStyle/index.ts
@@ -59,7 +59,6 @@ export type ProAliasToken = AliasToken & {
    * @example .ant
    */
   antCls: string;
-  hashId?: string;
 };
 
 export const resetComponent = (token: ProAliasToken): CSSObject => ({
@@ -111,9 +110,8 @@ export function useStyle(
     wrapSSR: useStyleRegister({ theme, token, hashId, path: [componentName] }, () =>
       styleFn({
         ...token,
-        hashId,
         antCls: '.' + getPrefixCls(),
-        proComponentsCls: hashId + proComponentsCls,
+        proComponentsCls,
       }),
     ),
     hashId,