+++ packages/card/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-experimental.21](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@2.0.0-experimental.20...@ant-design/pro-card@2.0.0-experimental.21) (2022-08-25)
+
+**Note:** Version bump only for package @ant-design/pro-card
+
+# [2.0.0-experimental.20](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@2.0.0-experimental.19...@ant-design/pro-card@2.0.0-experimental.20) (2022-08-25)
+
+**Note:** Version bump only for package @ant-design/pro-card
+
+# [2.0.0-experimental.19](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.20.19...@ant-design/pro-card@2.0.0-experimental.19) (2022-08-25)
+
+### Bug Fixes
+
+- card style fixed ([#5741](https://github.com/ant-design/pro-components/issues/5741)) ([4698978](https://github.com/ant-design/pro-components/commit/46989786151df482c3109240942716002adc3d44))
+
+### Features
+
+- add card tabs change ([#5761](https://github.com/ant-design/pro-components/issues/5761)) ([5504423](https://github.com/ant-design/pro-components/commit/55044237907b466991e91297f4fe623b185dc8df))
+- cssinjs ([#5596](https://github.com/ant-design/pro-components/issues/5596)) ([b82c7d2](https://github.com/ant-design/pro-components/commit/b82c7d2a1021506918947df56aaa66c3181187c7))
+
+# [2.0.0-experimental.18](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@2.0.0-experimental.17...@ant-design/pro-card@2.0.0-experimental.18) (2022-08-24)
+
+**Note:** Version bump only for package @ant-design/pro-card
+
 # [2.0.0-experimental.17](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@2.0.0-experimental.16...@ant-design/pro-card@2.0.0-experimental.17) (2022-08-24)
 
 **Note:** Version bump only for package @ant-design/pro-card